requirements.txt
Django==4.0.4
django-browser-reload==1.3.0
django-light==0.1.0.post3
django-taggit==2.1.0
django-tailwind==3.1.1
mysqlclient==2.1.0
Pillow==9.1.0
django-environ==0.8.1
python-dotenv==0.20.0
boto3==1.22.13
django-storages==1.12.3
django-imagekit==4.1.0
